requirements.txt
beautifulsoup4==4.10.0
python-dotenv==0.19.0
ipaddress==1.0.23
requests==2.26.0